[TASK] Adjust code to new vendor namespace and fix bugs
This change set contains all the simple replacements adjusting the
code to the new \TYPO3\* vendor namespace and the new syntax for
package keys.

Also fixes a few bugs and the unit tests of this package. In combination
with the new vendor namespaces code base of FLOW3 this kickstarter
should now be fully functional again ;-)

Change-Id: I92e4d071ec9ef3c5e7a614f326832f2369fcc4e7
Related: #27490